requirements.txt
langchain==0.3.26
langchain-core==0.3.71
langchain-community==0.3.27
langchain-google-genai==2.1.8
langchain-openai==0.3.28
langsmith==0.4.8
# 👇 DO NOT use `google-generativeai` when using `langchain-google-genai`
python-dotenv==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.7.0
tqdm
beautifulsoup4
requests
duckduckgo_search
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn
